Use double precision for KaxDuration output.

+++ src/mkvinfo.cpp
@@ -675,8 +675,8 @@ def_handle(info) {
     } else if (is_id(l2, KaxDuration)) {
       KaxDuration &duration = *static_cast<KaxDuration *>(l2);
       show_element(l2, 2, "Duration: %.3fs (" FMT_TIMECODEN ")",
-                   float(duration) * tc_scale / 1000000000.0,
-                   ARG_TIMECODEN(float(duration) * tc_scale));
+                   double(duration) * tc_scale / 1000000000.0,
+                   ARG_TIMECODEN(double(duration) * tc_scale));
 
     } else if (is_id(l2, KaxMuxingApp)) {
       KaxMuxingApp &muxingapp = *static_cast<KaxMuxingApp *>(l2);
